Switch to node-build
This switches asdf-nodejs to use node-build as the backend. This fixes
support for platforms that do not have binaries, and removes the
somewhat problematic GPG verification code.

Instead, we rely on the hand-curated list of definitions that are
maintained at the node-build repo. This also makes listing possible
versions of node much faster, and introduces support for io.js (if
anyone still needs it, I suppose) and other forks.