Add handling for missing shasum
On some platforms, including Fedora and derivatives, and Arch and
derivatives, the binary `shasum` does not exist.  Instead, sha256sum
should be used.

This patch checks for the existence of sha256sum, as well as shasum,
and uses whichever is present.  If none are present, a helpful error
message is printed out.

Tested on Fedora 28
  - Scenarios tested:

    -------------------------------
    | Sha256sum | shasum | Works? |
    |    Yes    |   No   |  Yes   |
    |    No     |   Yes  |  Yes   |
    |    Yes    |   Yes  |  Yes   |
    |    No     |   No   |  Yes   |
    -------------------------------