Fixes #85 - system npm fails to run via `asdf`
`.asdf/bin/private/asdf-exec` sets `$ASDF_INSTALL_PATH` to the result of
`find_install_path` in `.asdf/lib/utils.sh` - which, if version is
`"system"` (which it is for the system nodejs, of course), returns an
empty string. This then, is passed to
`.asdf/plugins/nodejs/bin/exec-env`, and, if `$NPM_CONFIG_PREFIX` is
`""`, it is set to `$ASDF_INSTALL_PATH/.npm` - resulting in `"/.npm"`.

In my shell w/o `asdf` enabled which can run `npm`, `$NPM_CONFIG_PREFIX`
is blank; therefore a possible patch would be to only set
`$NPM_CONFIG_PREFIX` if it is blank _and_ `$ASDF_INSTALL_PATH` is _not_.