install: Try to use ASDF_DATA_DIR
This allows the GPG keyrings to be installed into ${ASDF_DATA_DIR}
rather than ${ASDF_DIR}; this is useful if asdf itself may be replaced
wholesale (e.g. by a package manager).  Semantically, it also makes more
sense to put the keyring into the data directory, as it is not part of
the core asdf application.