Install most recent point release of inexact versions
This change will make it such that when installing a version like so:

`asdf install nodejs 16.10`

It would install 16.10.3, assuming 16.10.3 is the latest point release
at the time of installation.  This change allows asdf to work better
with legacy version specifiers such as what would be stored in `.nvmrc`.

For this to work properly, it would also (in the example above), symlink
16.10 to point to 16.10.3 (and likewise to link 16 to 16.10.3).  Such
symlinks are expected to point to the latest version installed that
satisfies the partial version number.

An uninstallation script has to be added so that symlinks aren't left
dangling at any point.

This fix addresses issue #295.